pyproject.toml
[tool.poetry]
name = "project_planner"
version = "0.1.0"
description = "An AI assistent for project planning"
authors = [
    "Shima Rashidi <shima.rashidi7@gmail.com>"
]
packages = [
    { include = "agents" },
]

[tool.poetry.dependencies]
python = "~3.11"
langgraph = "^0.2.0"
langchain_anthropic = "^0.1.0"
langchain_core = "^0.2.0"
langchain_openai = "^0.1.0"
tavily-python = "^0.3.0"
langchain_community = "^0.2.0"
pre-commit = "^4.0.1"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
